check-cccd/requirements.txt
fastapi
uvicorn[standard]
requests[socks]
beautifulsoup4
lxml
celery[redis]
redis
sqlalchemy
alembic
psycopg2-binary
pytest
pytest-httpx
respx
python-dotenv
cryptography
pydantic-settings
structlog
prometheus-client
sentry-sdk[fastapi]
pandas
openpyxl
